+++ spi/macronix.c
@@ -31,9 +31,9 @@
 //#define SPI_DEBUG
 
 #include <stdlib.h>
-#include "spi_flash.h"
+#include <spi/spi_flash.h>
+#include <spi/spi.h>
 #include "spi_flash_internal.h"
-#include "spi.h"
 
 /* MX25xx-specific commands */
 #define CMD_MX25XX_WREN       0x06	/* Write Enable */