+++ CHANGELOG.md
@@ -8,6 +8,16 @@ Releases 4.5.x and 4.6.x are based on mainline support submitted in
 
 ## [Unreleased]
 
+## [v4.6.4] - 2017-11-30
+### Added
+- hidden menu with security registers access for writing the custom serial
+  number
+- `Legacy serial console redirection` option removed
+- automated versioning
+
+### Changed
+- some code refactoring
+
 ## [v4.6.3] - 2017-10-30
 ### Added
 - Enable runtime configuration of UARTc/d, EHCI and mPCIe2 CLK in mainline
@@ -110,7 +120,8 @@ initial commit based on [coreboot_140908](http://pcengines.ch/tmp/coreboot_14090
 ### Fixed
 - used proper way to access extended SPI registers
 
-[Unreleased]: https://github.com/pcengines/sortbootorder/compare/v4.6.3...master
+[Unreleased]: https://github.com/pcengines/sortbootorder/compare/v4.6.4...master
+[v4.6.4]: https://github.com/pcengines/sortbootorder/compare/v4.6.3...v4.6.4
 [v4.6.3]: https://github.com/pcengines/sortbootorder/compare/v4.5.7...v4.6.3
 [v4.5.7]: https://github.com/pcengines/sortbootorder/compare/v4.5.6...v4.5.7
 [v4.5.6]: https://github.com/pcengines/sortbootorder/compare/v4.5.5...v4.5.6